feat: add auth_sessions table and related tests
- Updated schema.sql to include new auth_sessions table with relevant fields.
- Added profile_pic field to customers and users tables.
- Enhanced pyproject.toml and uv.lock with new dependencies: httpx, pytest, and python-multipart.
- Created tests for auth session management, including login, refresh token handling, and session revocation.

+++ app/models/tour_variant.py
@@ -1,15 +1,8 @@
 import uuid
-from sqlalchemy import Boolean
-from sqlalchemy import Date
-from sqlalchemy import ForeignKey
-from sqlalchemy import Integer
-from sqlalchemy import Numeric
-from sqlalchemy import String
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import Mapped
-from sqlalchemy.orm import mapped_column
-from sqlalchemy.orm import relationship
+from sqlalchemy import Boolean, Date, ForeignKey, Integer, Numeric, String, ForeignKey
+from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.models.base import ActiveEntity
+from decimal import Decimal
 
 
 class TourVariant(ActiveEntity):
@@ -60,7 +53,7 @@ class TourVariant(ActiveEntity):
         nullable=False,
     )
 
-    base_price: Mapped[float] = mapped_column(
+    base_price: Mapped[Decimal] = mapped_column(
         Numeric(10, 2),
         nullable=False,
     )
